server/tests/test_resource_management.py
"""Tests for GPU resource management concepts."""

from __future__ import annotations

import unittest

import pytest
from pydantic import ValidationError


class TestCleanupResources(unittest.TestCase):
    """Test cleanup_resources returns expected format."""

    def test_cleanup_result_format(self):
        """Cleanup should return dict with freed_mb and message keys."""
        # The actual function requires CUDA; test the expected contract
        result = {"freed_mb": 512.0, "message": "Freed 512.0 MB VRAM"}
        assert "freed_mb" in result
        assert "message" in result
        assert isinstance(result["freed_mb"], (int, float))
        assert result["freed_mb"] >= 0

    def test_cleanup_zero_freed(self):
        """Cleanup with no GPU should report 0 freed."""
        result = {"freed_mb": 0.0, "message": "Cleanup complete (no GPU)"}
        assert result["freed_mb"] == 0.0
        assert "message" in result


class TestModeTransitions(unittest.TestCase):
    """Test that mode transitions follow expected patterns."""

    def test_txt2img_does_not_require_source(self):
        """txt2img mode should not require source image."""
        from sddj.protocol import GenerateRequest, GenerationMode

        req = GenerateRequest(
            prompt="test",
            mode=GenerationMode.TXT2IMG,
        )
        assert req.source_image is None

    def test_img2img_requires_source(self):
        """img2img mode should require source image."""
        from sddj.protocol import GenerateRequest, GenerationMode

        with pytest.raises(ValidationError, match="source_image"):
            GenerateRequest(
                prompt="test",
                mode=GenerationMode.IMG2IMG,
            )

    def test_img2img_valid_with_source(self):
        """img2img mode should succeed when source image is provided."""
        from sddj.protocol import GenerateRequest, GenerationMode

        req = GenerateRequest(
            prompt="test",
            mode=GenerationMode.IMG2IMG,
            source_image="base64data",
        )
        assert req.source_image == "base64data"
        assert req.mode == GenerationMode.IMG2IMG
