server/pyproject.toml
[project]
name = "sddj-server"
version = "0.7.5"
description = "Local SOTA generation and animation server for Aseprite via Stable Diffusion"
requires-python = ">=3.11,<3.14"
license = "MIT"
dependencies = [
    # Web server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14",
    # Diffusion engine
    "torch>=2.4,<3.0",
    "diffusers>=0.31,<1.0",
    "transformers>=4.45,<6.0",
    "accelerate>=0.34,<2.0",
    "safetensors>=0.4,<1.0",
    "peft>=0.13,<1.0",
    # Speed optimizations
    "DeepCache>=0.1",
    "triton-windows>=3.2,<3.7; sys_platform == 'win32'",
    "triton>=3.2; sys_platform != 'win32'",
    # Post-processing + motion warp
    "opencv-python-headless>=4.8",
    "Pillow>=10.0,<13.0",
    "scikit-learn>=1.5",
    "scikit-image>=0.24",
    "scipy>=1.13",
    "numpy>=1.26,<3.0",
    "numba>=0.60",
    "huggingface_hub>=0.25",
    # Background removal
    "rembg>=2.0,<3.0",
    "onnxruntime>=1.17",
    # Config
    "pydantic-settings>=2.5",
    # Audio reactivity
    "librosa>=0.10,<1.0",
    "simpleeval>=1.0.5,<2.0",
    "soundfile>=0.12",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27"]
audio-stems = ["demucs>=4.0,<5.0"]

[project.scripts]
sddj-server = "sddj.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sddj"]

# ── uv: force PyTorch CUDA 12.8 builds ──────────────────────
[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]
